runtime/object: split array object ops
