+++ crates/perry-runtime/src/object/object_ops.rs
@@ -2,10 +2,6 @@
 //! `Object.fromEntries`/`groupBy`/`is`/`hasOwn`/`create`/`freeze`/`seal`/
 //! `defineProperty`/`getOwnPropertyDescriptor`/`getPrototypeOf`/... plus
 //! the `js_object_*` helpers backing them.
-//!
-//! Split out of `object.rs` (issue #1103). Pure relocation. The
-//! `globalThis` singleton subsystem stays in the parent module because
-//! it is also consumed by class/builtin-resolution code there.
 
 use super::*;
 
@@ -577,12 +573,7 @@ pub extern "C" fn js_object_has_own(obj_value: f64, key_value: f64) -> f64 {
     }
 }
 
-/// `Object.prototype.propertyIsEnumerable.call(obj, key)` (#2891) — true iff
-/// `key` is an OWN property of `obj` whose descriptor is enumerable. Inherited
-/// and absent keys return false. Nullish receivers throw `TypeError` per
-/// ToObject. Mirrors the ordinary-method branch in `native_call_method.rs`
-/// (which handles `obj.propertyIsEnumerable(key)`); this entry point is what
-/// the syntactic `.call` shapes lower to.
+/// `Object.prototype.propertyIsEnumerable.call(obj, key)` (#2891).
 #[no_mangle]
 pub extern "C" fn js_object_property_is_enumerable(obj_value: f64, key_value: f64) -> f64 {
     const TAG_TRUE: u64 = 0x7FFC_0000_0000_0004;
@@ -645,26 +636,8 @@ pub extern "C" fn js_object_property_is_enumerable(obj_value: f64, key_value: f6
             Ok(s) => s,
             Err(_) => return f64::from_bits(TAG_FALSE),
         };
-        if (obj as usize) >= crate::gc::GC_HEADER_SIZE + 0x1000 {
-            let gc_header =
-                (obj as *const u8).sub(crate::gc::GC_HEADER_SIZE) as *const crate::gc::GcHeader;
-            if (*gc_header).obj_type == crate::gc::GC_TYPE_ARRAY {
-                if key_name == "length" {
-                    return f64::from_bits(TAG_FALSE);
-                }
-                let arr = obj as *const crate::array::ArrayHeader;
-                if super::has_own_helpers::array_own_key_present(arr, key_str) {
-                    let enumerable = if super::canonical_array_index(key_name).is_some() {
-                        true
-                    } else {
-                        super::get_property_attrs(obj as usize, key_name)
-                            .map(|attrs| attrs.enumerable())
-                            .unwrap_or(true)
-                    };
-                    return f64::from_bits(if enumerable { TAG_TRUE } else { TAG_FALSE });
-                }
-                return f64::from_bits(TAG_FALSE);
-            }
+        if let Some(result) = super::array_property_is_enumerable(obj, key_str, key_name) {
+            return result;
         }
         if !is_valid_obj_ptr(obj as *const u8) {
             return f64::from_bits(TAG_FALSE);
@@ -913,71 +886,14 @@ pub extern "C" fn js_object_define_property(
             let name_bytes = std::slice::from_raw_parts(name_ptr, name_len);
             std::str::from_utf8(name_bytes).ok().map(|s| s.to_string())
         };
-        if (obj as usize) >= crate::gc::GC_HEADER_SIZE + 0x1000 {
-            let gc_header =
-                (obj as *const u8).sub(crate::gc::GC_HEADER_SIZE) as *const crate::gc::GcHeader;
-            if (*gc_header).obj_type == crate::gc::GC_TYPE_ARRAY {
-                let Some(ref key_name) = key_rust else {
-                    return obj_value;
-                };
-                let desc_ptr = extract_obj_ptr(descriptor_value);
-                if desc_ptr.is_null() {
-                    return obj_value;
-                }
-                let value_key = crate::string::js_string_from_bytes(b"value".as_ptr(), 5);
-                let has_value = own_key_present(desc_ptr, value_key);
-                let value_field =
-                    js_object_get_field_by_name(desc_ptr as *const ObjectHeader, value_key);
-                let value = if has_value {
-                    f64::from_bits(value_field.bits())
-                } else {
-                    f64::from_bits(crate::value::TAG_UNDEFINED)
-                };
-
-                if key_name == "length" {
-                    if has_value {
-                        crate::array::js_array_set_length(
-                            obj as *mut crate::array::ArrayHeader,
-                            value,
-                        );
-                    }
-                    return obj_value;
-                }
-                if let Some(index) = super::canonical_array_index(key_name) {
-                    if has_value {
-                        crate::array::js_array_set_f64_extend(
-                            obj as *mut crate::array::ArrayHeader,
-                            index,
-                            value,
-                        );
-                    }
-                    return obj_value;
-                }
-
-                crate::array::array_named_property_set(
-                    obj as *mut crate::array::ArrayHeader,
-                    key_str,
-                    value,
-                );
-
-                let read_bool = |name: &[u8]| -> Option<bool> {
-                    let k = crate::string::js_string_from_bytes(name.as_ptr(), name.len() as u32);
-                    if !own_key_present(desc_ptr, k) {
-                        return None;
-                    }
-                    let v = js_object_get_field_by_name(desc_ptr as *const ObjectHeader, k);
-                    Some(crate::value::js_is_truthy(f64::from_bits(v.bits())) != 0)
-                };
-                let writable = read_bool(b"writable").unwrap_or(false);
-                let enumerable = read_bool(b"enumerable").unwrap_or(false);
-                let configurable = read_bool(b"configurable").unwrap_or(false);
-                set_property_attrs(
-                    obj as usize,
-                    key_name.clone(),
-                    PropertyAttrs::new(writable, enumerable, configurable),
-                );
-                return obj_value;
-            }
+        if let Some(result) = super::define_array_property(
+            obj,
+            obj_value,
+            key_str,
+            key_rust.as_deref(),
+            descriptor_value,
+        ) {
+            return result;
         }
         // #2843: enforce frozen / sealed / non-extensible invariants BEFORE any
         // mutation, so a rejected definition leaves the object untouched and the
@@ -1432,18 +1348,6 @@ pub extern "C" fn js_get_global_this_builtin_value(name_ptr: *const u8, name_len
     f64::from_bits(bits)
 }
 
-fn builtin_constructor_prototype_value(name: &[u8]) -> Option<f64> {
-    let ctor = js_get_global_this_builtin_value(name.as_ptr(), name.len());
-    let ctor_value = crate::value::JSValue::from_bits(ctor.to_bits());
-    if !ctor_value.is_pointer() {
-        return None;
-    }
-    let ctor_ptr = ctor_value.as_pointer::<u8>() as usize;
-    let proto = crate::closure::closure_get_dynamic_prop(ctor_ptr, "prototype");
-    let proto_value = crate::value::JSValue::from_bits(proto.to_bits());
-    proto_value.is_pointer().then_some(proto)
-}
-
 /// Object.create(proto) — create empty object. Perry ignores prototype; Object.create(null) returns {}.
 #[no_mangle]
 pub extern "C" fn js_object_create(proto_value: f64) -> f64 {
@@ -1635,14 +1539,8 @@ pub extern "C" fn js_object_get_prototype_of(obj_value: f64) -> f64 {
                     }
                 }
                 if (*gc).obj_type == crate::gc::GC_TYPE_ARRAY {
-                    if let Some(array_proto) = builtin_constructor_prototype_value(b"Array") {
-                        let proto_addr = crate::value::js_nanbox_get_pointer(array_proto) as u64;
-                        if proto_addr != raw_addr {
-                            return array_proto;
-                        }
-                    }
-                    if let Some(object_proto) = builtin_constructor_prototype_value(b"Object") {
-                        return object_proto;
+                    if let Some(proto) = super::array_get_prototype_of_addr(raw_addr as usize) {
+                        return proto;
                     }
                 }
                 // #489 / #2145: a function/constructor receiver has no
@@ -1691,14 +1589,8 @@ pub extern "C" fn js_object_get_prototype_of(obj_value: f64) -> f64 {
                     }
                 }
                 if (*gc).obj_type == crate::gc::GC_TYPE_ARRAY {
-                    if let Some(array_proto) = builtin_constructor_prototype_value(b"Array") {
-                        let proto_addr = crate::value::js_nanbox_get_pointer(array_proto);
-                        if proto_addr != bits as i64 {
-                            return array_proto;
-                        }
-                    }
-                    if let Some(object_proto) = builtin_constructor_prototype_value(b"Object") {
-                        return object_proto;
+                    if let Some(proto) = super::array_get_prototype_of_addr(bits as usize) {
+                        return proto;
                     }
                 }
                 // #489 / #2145: function/constructor receiver — see the
